more changes

+++ scala/redbubble2/src/main/scala/redbubble.scala
@@ -3,8 +3,8 @@
   */
 
 import java.io.{FileWriter, BufferedWriter, File}
-import scala.util.{Try,Success,Failure}
-import scala.io.Source
+import java.nio.file.{Paths, Files}
+import scala.util.{Try, Success, Failure}
 import com.typesafe.config.ConfigFactory
 import scala.xml._
 
@@ -13,12 +13,6 @@ case class Thumbnail(size: String, url: String)
 case class MakeAndThumbNail(make: String, thumbnail: Thumbnail)
 case class ModelAndThumbNail(model: String, thumbnail: Thumbnail)
 case class AllMakeModelThumbNails(make: String, model: String, thumbnail: Thumbnail)
-
-abstract class Cameras {
-  val cameraMake: String
-  val cameraModels: Seq[String]
-  val thumbNailURLs: Seq[Thumbnail]
-}
 
 class Redbubble(node: scala.xml.Elem) {
 
@@ -29,104 +23,98 @@ class Redbubble(node: scala.xml.Elem) {
   val Url_style = "<style type=\"text/css\">\nnav { \nmargin: 10px;}\n</style>\n</head>\n<body>\n<header>"
   val Url_end = "</header>\n</body></html>"
 
-  val BaseDir = System.getProperty("user.dir")
+  val _BaseDir = ConfigFactory.load().getConfig("redbubble").getString("base_dir")
+  val BaseDir = if (_BaseDir.trim.isEmpty) System.getProperty("user.dir") else _BaseDir
   val TargetHtml = ConfigFactory.load().getConfig("redbubble").getString("html_path")
-  val IndexPath:  String = s"$BaseDir$TargetHtml/index.html"
+  val IndexPath: String = s"$BaseDir$TargetHtml"
   val CameraMakePath = ConfigFactory.load().getConfig("redbubble").getString("camera_make_path")
   val CameraModelPath = ConfigFactory.load().getConfig("redbubble").getString("camera_model_path")
 
-  val Url_index:  String = "<nav>\n<a href=\"" + IndexPath + "\">Index</a>\n</nav>"
+  val Url_index: String = "<nav>\n<a href=\"" + IndexPath + "/index.html"+ "\">Index</a>\n</nav>"
 
   val DefaultImageHeight = ConfigFactory.load().getConfig("redbubble").getString("thumbnail_image_height")
   val DefaultImageWidth = ConfigFactory.load().getConfig("redbubble").getString("thumbnail_image_width")
   val AllPictureSize = ConfigFactory.load().getConfig("redbubble").getString("all_picture_size")
   val NumberOfWorktoDisplay = ConfigFactory.load().getConfig("redbubble").getInt("number_of_work_to_display")
+  //**************
 
-  def fromXML(make: String, model: Seq[String], thumbNails: Seq[Thumbnail]): Cameras =
-    new Cameras {
-      val cameraMake: String = make
-      val cameraModels: Seq[String] = model
-      val thumbNailURLs: Seq[Thumbnail] = thumbNails
+  def createDirectoryStructure: Boolean = {
+    val htmlDir: File = new File(BaseDir + TargetHtml)
+    val cameraMakeDir: File = new File(BaseDir + TargetHtml + CameraMakePath)
+    val cameraModelDir: File = new File(BaseDir + TargetHtml + CameraModelPath)
+
+    val checkBasePathExists: Boolean = Files.exists(Paths.get(BaseDir + TargetHtml))
+
+    if (!checkBasePathExists) {
+      htmlDir.mkdir()
+      cameraMakeDir.mkdir()
+      cameraModelDir.mkdir()
+      true
     }
-
-  def cameraMap: Map[String, Set[String]] = {
-
-    //maps every model to it's make, trimming works that don't have a camera defined
-    //Map(D-LUX 3 -> LEICA, DMC-FZ30 -> Panasonic, ......
-    val modelMakeMap = (
-      for {n <- node.child
-           if (!((n \\ "make").text.trim.isEmpty))
-      } yield {
-        (n \\ "model").text -> (n \\ "make").text
-      }
-      ).toMap
-
-    //reverse the mapping above, and groups models to makes
-    //Map(LEICA -> Set(D-LUX 3),FUJI PHOTO FILM CO., LTD. -> Set(SLP1000SE), NIKON CORPORATION -> Set(NIKON D80),
-    (modelMakeMap.groupBy(_._2).mapValues(_.keys.toSet))
-  }
-
-  private def processXML(cameraMaps: Map[String, Set[String]]): Seq[Cameras] = {
-
-    (for {(cameraMake, cameraModel) <- cameraMaps}
-      yield {
-        fromXML(cameraMake, getAllCameraModels(cameraMake), getCameraMakeThumbNails(cameraMake))
-      }).toSeq
-  }
-
-  //get all thumbNails from file with cameras. Where a cameraMake doesn't exist, substitute empty string with 'None'
-  def getAllMakeAndThumbnails: Seq[MakeAndThumbNail] = {
-    for {n <- node.child
-         urls <- (n \\ "url")
-    } yield {
-      val make = if( (n \\ "make").text.trim.isEmpty ) "None" else (n \\ "make").text.trim
-
-      MakeAndThumbNail(make , (Thumbnail((urls \ "@type").mkString, urls.text)))
+    else {
+      Console.err.println(s"Error: Could not create directory structure. Paths already exist")
+      false
     }
   }
 
   //case class AllMakeModelThumbNails(make: String, model: String, thumbnail: Thumbnail)
-  val allMakeModelThumbNails: Seq[AllMakeModelThumbNails] = {
+  //contains all mapping from make -> model(s) -> thumbnail(s)
+  //**************
+  private val allMakeModelThumbNails: Seq[AllMakeModelThumbNails] = {
     for {n <- node.child
          urls <- (n \\ "url")
     } yield {
-      val make = if( (n \\ "make").text.trim.isEmpty ) "None" else (n \\ "make").text.trim
+      val make = if ((n \\ "make").text.trim.isEmpty) "None" else (n \\ "make").text.trim
 
-      AllMakeModelThumbNails(make , (n \\ "model").text, (Thumbnail((urls \ "@type").mkString, urls.text)))
+      AllMakeModelThumbNails(make, (n \\ "model").text, (Thumbnail((urls \ "@type").mkString, urls.text)))
+    }
+  }
+
+  //get all thumbNails from file with cameras. Where a cameraMake doesn't exist, substitute empty string with 'None'
+  private def allMakeAndThumbnails: Seq[MakeAndThumbNail] = {
+    for {n <- node.child
+         urls <- (n \\ "url")
+    } yield {
+      val make = if ((n \\ "make").text.trim.isEmpty) "None" else (n \\ "make").text.trim
+
+      MakeAndThumbNail(make, (Thumbnail((urls \ "@type").mkString, urls.text)))
     }
   }
 
   //get ALL camera models that exist in the file, even duplicates
-  private def getAllCameraModels(cameraMake: String): Seq[String] =
-     allMakeModelThumbNails.filter(_.make == cameraMake).map(_.model)
+  def allCameraModels(cameraMake: String): Seq[String] =
+    allMakeModelThumbNails.filter(_.make == cameraMake.trim).map(_.model)
 
   //get thumbNails specific to a camera make
-  private def getCameraMakeThumbNails(cameraMake: String): Seq[Thumbnail] =
-    getAllMakeAndThumbnails.filter(_.make == cameraMake).map(_.thumbnail)
+  def cameraMakeThumbNails(cameraMake: String): Seq[Thumbnail] =
+    allMakeAndThumbnails.filter(_.make == cameraMake).map(_.thumbnail)
 
-  private def getCameraModelThumbNails(cameraModel: String): Seq[Thumbnail] =
+  def cameraModelThumbNails(cameraModel: String): Seq[Thumbnail] =
     allMakeModelThumbNails.filter(_.model == cameraModel).map(_.thumbnail)
 
-  private val allCameraMakes: Seq[String] =
+  val allCameraMakes: Seq[String] =
     allMakeModelThumbNails.map(_.make).distinct
 
   //needed because not all works contain a camera make. requirement is to grab first 10 thumbnail-works, but doesn't
   //mention whether a camera created the photo
-  private val getIndexThumbnails: Seq[Thumbnail] =
-    getAllMakeAndThumbnails.map(_.thumbnail)
+  val allIndexThumbnails: Seq[Thumbnail] =
+    allMakeAndThumbnails.map(_.thumbnail)
 
   private def htmlAhref(toPrint: String, path: String): String = {
     ("<a href=\"" + BaseDir + TargetHtml + path + "/" +
       toPrint + ".html\">" + toPrint + "</a> | \n").mkString("")
   }
+
   private def htmlTitle(toPrint: String): String = {
     s"<title>$toPrint</title>"
   }
+
   private def htmlHeader(toPrint: String): String = {
     s"<h1>$toPrint</h1>"
   }
+
   private def htmlIMG(url: String): String = {
-    "<img height=\""+ getThumbnailHeight(url) + "\"width=\"" + getThumbnailWidth(url) + "\"src=\"" + (url)+ "\">\n"
+    "<img height=\"" + getThumbnailHeight(url) + "\"width=\"" + getThumbnailWidth(url) + "\"src=\"" + (url) + "\">\n"
   }
 
   private def getThumbnailHeight(url: String) = {
@@ -140,6 +128,7 @@ class Redbubble(node: scala.xml.Elem) {
       case Failure(f) => DefaultImageHeight
     }
   }
+
   private def getThumbnailWidth(url: String) = {
     val getWidth: Try[String] = {
       Try(url.split(",")(1).split('x')(0))
@@ -151,91 +140,106 @@ class Redbubble(node: scala.xml.Elem) {
     }
   }
 
-  private def getThumbNails(whichPage: String, generic: String) = {
-
-  }
-
   private def applyConstraints(tb: Seq[Thumbnail]): String = {
-                                      tb.filter(_.size == AllPictureSize)
-                                      .take(NumberOfWorktoDisplay)
-                                      .map(x => htmlIMG(x.url))
-                                      .mkString("")
+    tb.filter(_.size == AllPictureSize)
+      .take(NumberOfWorktoDisplay)
+      .map(x => htmlIMG(x.url))
+      .mkString("")
   }
 
-  def printHTMLindex(allWorks: Seq[Cameras]) = {
+  private def checkPathExists(path: String): Boolean = {
+    Files.exists(Paths.get(path))
+  }
 
-    val title:  String = htmlTitle("Navigation for Works")
+  //**************
+  //methods to print to HTML
+  //**************
+  def printHTML = {
+
+    val title: String = htmlTitle("Navigation for Works")
     val header: String = htmlHeader("Camera Makes")
 
     val cameraMakes = (
-      for{ cameraMake <- allCameraMakes if(cameraMake != "None")
+      for {cameraMake <- allCameraMakes if (cameraMake != "None")
       } yield {
-          printCameraMakeWebPages(cameraMake)
-          htmlAhref(cameraMake, CameraMakePath)
+        printCameraMakeWebPages(cameraMake)
+        htmlAhref(cameraMake, CameraMakePath)
       }).mkString("")
 
     val navCameraMakes = s"<nav>$cameraMakes</nav>"
 
-    val navThumbNails = applyConstraints(getIndexThumbnails)
+    val navThumbNails = applyConstraints(allIndexThumbnails)
 
     val html = Url_head + title + Url_style + header + navCameraMakes + navThumbNails + Url_end
 
-    writeFile(IndexPath, html)
+    writeFile(IndexPath, "/index.html", html)
   }
 
   def printCameraMakeWebPages(cameraMake: String) = {
 
     //*************
-    val title:        String = "<title>Navigation for Works</title>"
-    val header:       String = s"<h1>$cameraMake</h1>"
+    val title: String = "<title>Navigation for Works</title>"
+    val header: String = s"<h1>$cameraMake</h1>"
 
-    val cameraModels = (for(model <- getAllCameraModels(cameraMake).distinct) yield {
+    val cameraModels = (for (model <- allCameraModels(cameraMake).distinct) yield {
       printCameraModelWebPages(model)
       htmlAhref(model, CameraModelPath)
     }).mkString("")
 
     val navCameraModels = s"<nav>$cameraModels</nav>"
 
-    val navCameraMakeThumbNails = applyConstraints(getCameraMakeThumbNails(cameraMake))
+    val navCameraMakeThumbNails = applyConstraints(cameraMakeThumbNails(cameraMake))
 
     val html = Url_head + title + Url_style + header + Url_index + navCameraModels + navCameraMakeThumbNails + Url_end
 
-    writeFile(s"$BaseDir$TargetHtml$CameraMakePath/$cameraMake.html", html.toString)
+    writeFile(s"$BaseDir$TargetHtml$CameraMakePath", s"/$cameraMake.html", html.toString)
 
   }
 
   def printCameraModelWebPages(cameraModel: String) = {
 
     //*************
-    val title:        String = "<title>Navigation for Works</title>"
-    val header:       String = s"<h1>$cameraModel</h1>"
+    val title: String = "<title>Navigation for Works</title>"
+    val header: String = s"<h1>$cameraModel</h1>"
 
-    val modelThumbNails: String = applyConstraints(getCameraModelThumbNails(cameraModel))
+    val modelThumbNails: String = applyConstraints(cameraModelThumbNails(cameraModel))
 
     val html: String = Url_head + title + Url_style + header + Url_index + modelThumbNails + Url_end
 
-    writeFile(s"$BaseDir$TargetHtml$CameraModelPath/$cameraModel.html", html.toString)
+    writeFile(s"$BaseDir$TargetHtml$CameraModelPath", s"/$cameraModel.html", html.toString)
 
   }
 
-  def writeFile(path: String, stringToWrite: String) = {
-    val file = new File(s"$path")
-    val bw = new BufferedWriter(new FileWriter(file))
-    bw.write(stringToWrite)
-    bw.close()
+  def writeFile(path: String, fileName: String, stringToWrite: String): Boolean = {
+
+    if (checkPathExists(path)) {
+      val file: File = new File(s"$path$fileName")
+      val bw = new BufferedWriter(new FileWriter(file))
+      bw.write(stringToWrite)
+      bw.close()
+      true
+    } else {
+      Console.err.println(s"Error: $path does not exist")
+      false
+    }
   }
+
 }
 
+//***************
+//Companion object to class Redbubble.class
+//First we read in the XML file, load it via class scala.xml.XML and use that to process and print HTML
+//***************
 object Redbubble {
 
   def getFileFromUrl(file: String): String = {
     val readAPIFile: Try[List[String]] = {
-      Try(Source.fromURL(file).getLines.toList)
+      Try(scala.io.Source.fromURL(file).getLines.toList)
     }
 
     readAPIFile match {
       case Success(lines) => lines.mkString("")
-      case Failure(f) => s"NOT FOUND: $file"
+      case Failure(f) => s"Error: NOT FOUND: $file\n$f"
     }
   }
 
@@ -243,17 +247,17 @@ object Redbubble {
 
     println("Hello, world! ")
 
-    val url = ConfigFactory.load().getConfig("redbubble").getString("API_URL")
-    val node = XML.loadString(getFileFromUrl(url))
+    val Url = ConfigFactory.load().getConfig("redbubble").getString("API_URL")
+    val loadFile = getFileFromUrl(Url)
 
-    val rb = new Redbubble(node)
-
-    //println(s"In it's glory: ${rb.getAllMakeModelThumbNails}")
-
-    //translates the entire XML to the abstract class Cameras for easy access when creating HTML
-    val allWorks: Seq[Cameras] = rb.processXML(rb.cameraMap)
-
-    rb.printHTMLindex(allWorks)
+    if(!loadFile.contains("Error") && loadFile.contains("xml")) {
+      val node = XML.loadString(getFileFromUrl(Url))
+      val rb = new Redbubble(node)
+      rb.createDirectoryStructure
+      rb.printHTML
+    }
+    else
+      Console.err.println(s"****\nGracefully aborting program. Could not find $Url\n****")
 
   }
 }
